backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-rest-auth==0.9.3
djangorestframework==3.11.0
pytz==2021.1
PyYAML==5.4.1
six==1.11.0
sqlparse==0.4.1
uritemplate==3.0.1
